refactor(ucmc-web): address PR #55 review comments
- Move UnauthorizedError from principal.server.ts (which transitively
  imports Drizzle + KV at module scope) into a new lightweight
  errors.server.ts module. Consumers — including the
  /api/account/export route handler and exportMyDataAction — now
  import the sentinel without dragging in the heavier server deps.
  Same single-source-of-truth shape, smaller transitive cost.

- Drop the parenthetical Turnstile mention from the privacy "what we
  explicitly do not collect" bullet. The bullet is now strictly about
  what is NOT collected (third-party analytics, ad-tech identifiers,
  cross-site tracking); Turnstile's passive bot-screening signals are
  disclosed in the processors section below, which is the right place
  for it. Resolves the internal contradiction Copilot flagged.
